pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "operon-guard"
version = "0.1.0"
description = "Trust verification for AI agents — behavior specs, race-condition detection, safety rails, and trust scoring."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "Operon OS", email = "hello@operonos.com" }]
keywords = ["ai", "agents", "testing", "verification", "safety", "trust"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4"]

[project.scripts]
operon-guard = "operon_guard.cli:main"

[project.urls]
Homepage = "https://github.com/operonos/operon-guard"
Documentation = "https://docs.operonos.com/guard"
Repository = "https://github.com/operonos/operon-guard"
Issues = "https://github.com/operonos/operon-guard/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/operon_guard"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
